+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 * `enforce_https` config option forces the servers to reject non-ssl requests
 * `http_scheme` config option can be used to enforce HTTPS urls if proper proxy headers are missing
+* `/admin/debug` route for printing out greenlets
 ### Removed
 * JS Client code was moved to npm and separate repository
 ### Changed